requirements.txt
# Core RAG Pipeline
chromadb==1.5.0
langchain==1.2.10
langchain-community==0.4.1
langchain-core==1.2.12

# Document Processing
pypdf==6.7.0
python-docx==1.2.0
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==12.1.1

# Embeddings & Vector Search
sentence-transformers==5.2.2

# LLM Integration
requests==2.32.5

# Web Interface
streamlit==1.54.0

# Data Processing
pandas==2.3.3
numpy==2.4.2

# Utilities
python-dateutil==2.9.0.post0
tqdm==4.67.3
